code refactored

+++ src/main/java/com/etiya/rentACarSpring/businnes/concretes/CarMaintenanceManager.java
@@ -4,6 +4,7 @@ import java.util.Date;
 import java.util.List;
 import java.util.stream.Collectors;
 
+import com.etiya.rentACarSpring.businnes.abstracts.CarService;
 import com.etiya.rentACarSpring.businnes.abstracts.message.LanguageWordService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.core.env.Environment;
@@ -35,16 +36,19 @@ public class CarMaintenanceManager implements CarMaintenanceService {
     private RentalService rentalService;
     private Environment environment;
     private LanguageWordService languageWordService;
+    private CarService carService;
 
     @Autowired
     public CarMaintenanceManager(CarMaintenanceDao carMaintenanceDao, ModelMapperService modelMapperService,
-                                 RentalService rentalService, Environment environment,LanguageWordService languageWordService) {
+                                 RentalService rentalService, Environment environment,LanguageWordService languageWordService,
+                                 CarService carService) {
         super();
         this.carMaintenanceDao = carMaintenanceDao;
         this.modelMapperService = modelMapperService;
         this.rentalService = rentalService;
         this.environment = environment;
         this.languageWordService = languageWordService;
+        this.carService = carService;
     }
 
     @Override
@@ -54,7 +58,7 @@ public class CarMaintenanceManager implements CarMaintenanceService {
                         carMaintenance -> modelMapperService.forDto().map(carMaintenance, CarMaintenanceSearchListDto.class))
                 .collect(Collectors.toList());
 
-        return new SuccesDataResult<List<CarMaintenanceSearchListDto>>(response, languageWordService.getByLanguageAndKeyId(Messages.CarMaintenanceListed,Integer.parseInt(environment.getProperty("language"))));
+        return new SuccesDataResult<List<CarMaintenanceSearchListDto>>(response, languageWordService.getByLanguageAndKeyId(Messages.CarMaintenanceListed));
     }
 
     @Override
@@ -68,13 +72,16 @@ public class CarMaintenanceManager implements CarMaintenanceService {
         CarMaintenance carMaintenance = modelMapperService.forRequest().map(createCarMaintenanceRequest,
                 CarMaintenance.class);
         this.carMaintenanceDao.save(carMaintenance);
-        return new SuccesResult(languageWordService.getByLanguageAndKeyId(Messages.CarMaintenanceAdded,Integer.parseInt(environment.getProperty("language"))));
+        return new SuccesResult(languageWordService.getByLanguageAndKeyId(Messages.CarMaintenanceAdded));
     }
 
     @Override
     public Result update(UpdateCarMaintenanceRequest updateCarMaintenanceRequest) {
         CarMaintenance carMaintenance=this.carMaintenanceDao.getById(updateCarMaintenanceRequest.getCarMaintenanceId());
-        Result result = BusinnessRules.run(checkDate(carMaintenance.getMaintananceDate(),updateCarMaintenanceRequest.getReturnDate()));
+        Result result = BusinnessRules.run(checkDate(carMaintenance.getMaintananceDate(),updateCarMaintenanceRequest.getReturnDate()),
+                carService.checkCarExistsInGallery(updateCarMaintenanceRequest.getCarId()),
+                checkIfCarMaintenanceExists(updateCarMaintenanceRequest.getCarMaintenanceId()));
+
         if (result != null) {
             return result;
         }
@@ -84,13 +91,19 @@ public class CarMaintenanceManager implements CarMaintenanceService {
         updatedCarMaintenance.setCar(carMaintenance.getCar());
         updatedCarMaintenance.setMaintananceDate(carMaintenance.getMaintananceDate());
         this.carMaintenanceDao.save(updatedCarMaintenance);
-        return new SuccesResult(languageWordService.getByLanguageAndKeyId(Messages.CarMaintenanceUpdated,Integer.parseInt(environment.getProperty("language"))));
+        return new SuccesResult(languageWordService.getByLanguageAndKeyId(Messages.CarMaintenanceUpdated));
     }
 
     @Override
     public Result delete(DeleteCarMaintenanceRequest deleteCarMaintenanceRequest) {
+        Result result = BusinnessRules.run(
+                checkIfCarMaintenanceExists(deleteCarMaintenanceRequest.getCarMaintenanceId()));
+
+        if (result != null) {
+            return result;
+        }
         this.carMaintenanceDao.deleteById(deleteCarMaintenanceRequest.getCarMaintenanceId());
-        return new SuccesResult(languageWordService.getByLanguageAndKeyId(Messages.CarMaintenanceDeleted,Integer.parseInt(environment.getProperty("language"))));
+        return new SuccesResult(languageWordService.getByLanguageAndKeyId(Messages.CarMaintenanceDeleted));
     }
 
     @Override
@@ -104,7 +117,7 @@ public class CarMaintenanceManager implements CarMaintenanceService {
         if (result != null) {
             for (CarMaintenance carMaintenances : this.carMaintenanceDao.getByCar_CarId(carId)) {
                 if (carMaintenances.getReturnDate() == null) {
-                    return new ErrorResult(languageWordService.getByLanguageAndKeyId(Messages.CarIsAlreadyOnMaintenance,Integer.parseInt(environment.getProperty("language"))));
+                    return new ErrorResult(languageWordService.getByLanguageAndKeyId(Messages.CarIsAlreadyOnMaintenance));
                 }
             }
         }
@@ -114,7 +127,7 @@ public class CarMaintenanceManager implements CarMaintenanceService {
     private Result checkIfCarIsRentedNow(int carId) {
         Result isCarReturned = rentalService.checkCarRentalStatus(carId);
         if (!isCarReturned.isSuccess()) {
-            return new ErrorResult(languageWordService.getByLanguageAndKeyId(Messages.CarIsOnRent,Integer.parseInt(environment.getProperty("language"))));
+            return new ErrorResult(languageWordService.getByLanguageAndKeyId(Messages.CarIsOnRent));
         }
         return new SuccesResult();
     }
@@ -123,7 +136,13 @@ public class CarMaintenanceManager implements CarMaintenanceService {
         if (maintananceDate.compareTo(returnDate)<0){
             return new SuccesResult();
         }
-        return new ErrorResult(languageWordService.getByLanguageAndKeyId(Messages.InvalidReturnMaintenanceDate,Integer.parseInt(environment.getProperty("language"))));
+        return new ErrorResult(languageWordService.getByLanguageAndKeyId(Messages.InvalidReturnMaintenanceDate));
     }
 
+    private Result checkIfCarMaintenanceExists(int maintenanceId) {
+        if (!this.carMaintenanceDao.existsById(maintenanceId)) {
+            return new ErrorResult(languageWordService.getByLanguageAndKeyId(Messages.CarMaintenanceNotFound));
+        }
+        return new SuccesResult();
+    }
 }
