+++ src/main/java/com/etiya/rentACarSpring/businnes/concretes/CarManager.java
@@ -3,9 +3,13 @@ package com.etiya.rentACarSpring.businnes.concretes;
 import java.util.List;
 import java.util.stream.Collectors;
 
+import com.etiya.rentACarSpring.businnes.abstracts.BrandService;
+import com.etiya.rentACarSpring.businnes.abstracts.CityService;
+import com.etiya.rentACarSpring.businnes.abstracts.ColorService;
 import com.etiya.rentACarSpring.businnes.abstracts.message.LanguageService;
 import com.etiya.rentACarSpring.businnes.abstracts.message.LanguageWordService;
 import com.etiya.rentACarSpring.businnes.constants.Messages;
+import com.etiya.rentACarSpring.core.utilities.businnessRules.BusinnessRules;
 import com.etiya.rentACarSpring.core.utilities.results.*;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.core.env.Environment;
@@ -31,16 +35,22 @@ public class CarManager implements CarService {
     private findexScoreService findexScoreService;
     private Environment environment;
     private LanguageWordService languageWordService;
+    private CityService cityService;
+    private BrandService brandService;
+    private ColorService colorService;
 
     @Autowired
     public CarManager(CarDao carDao, ModelMapperService modelMapperService, findexScoreService findexScoreService, Environment environment
-            ,LanguageWordService languageWordService) {
+            ,LanguageWordService languageWordService,CityService cityService, BrandService brandService, ColorService colorService) {
         super();
         this.carDao = carDao;
         this.modelMapperService = modelMapperService;
         this.findexScoreService = findexScoreService;
         this.environment = environment;
         this.languageWordService = languageWordService;
+        this.cityService=cityService;
+        this.colorService=colorService;
+        this.brandService=brandService;
 
     }
 
@@ -50,30 +60,49 @@ public class CarManager implements CarService {
         List<CarSearchListDto> response = result.stream()
                 .map(car -> modelMapperService.forDto().map(car, CarSearchListDto.class)).collect(Collectors.toList());
 
-        return new SuccesDataResult<List<CarSearchListDto>>(response, languageWordService.getByLanguageAndKeyId(Messages.CarListed,Integer.parseInt(environment.getProperty("language"))));
+        return new SuccesDataResult<List<CarSearchListDto>>(response, languageWordService.getByLanguageAndKeyId(Messages.CarListed));
     }
 
     @Override
     public Result save(CreateCarRequest createCarRequest) {
+        Result result = BusinnessRules.run( cityService.checkIfCityExists(createCarRequest.getCityId()),
+                colorService.checkIfColorExists(createCarRequest.getColorId()),
+                brandService.checkIfBrandExists(createCarRequest.getBrandId())
+        );
+        if (result != null) {
+            return result;
+        }
         Car car = modelMapperService.forRequest().map(createCarRequest, Car.class);
         car.setFindexScore(findexScoreService.sendCarFindexScore());
         this.carDao.save(car);
-        return new SuccesResult(languageWordService.getByLanguageAndKeyId(Messages.CarAdded,Integer.parseInt(environment.getProperty("language"))));
+        return new SuccesResult(languageWordService.getByLanguageAndKeyId(Messages.CarAdded));
     }
 
     @Override
     public Result update(UpdateCarRequest updateCarRequest) {
+        Result result = BusinnessRules.run( checkCarExistsInGallery(updateCarRequest.getCarId()),
+                cityService.checkIfCityExists(updateCarRequest.getCityId()),
+                colorService.checkIfColorExists(updateCarRequest.getColorId()),
+                brandService.checkIfBrandExists(updateCarRequest.getBrandId())
+        );
+        if (result != null) {
+            return result;
+        }
         Car car = modelMapperService.forRequest().map(updateCarRequest, Car.class);
+        car.setFindexScore(this.carDao.getById(updateCarRequest.getCarId()).getFindexScore());
         this.carDao.save(car);
-        return new SuccesResult(languageWordService.getByLanguageAndKeyId(Messages.CarUpdated,Integer.parseInt(environment.getProperty("language"))));
+        return new SuccesResult(languageWordService.getByLanguageAndKeyId(Messages.CarUpdated));
     }
 
     @Override
     public Result delete(DeleteCarRequest deleteCarRequest) {
-
+        Result result = BusinnessRules.run( checkCarExistsInGallery(deleteCarRequest.getCarId())
+        );
+        if (result != null) {
+            return result;
+        }
         this.carDao.deleteById(deleteCarRequest.getCarId());
-        return new SuccesResult(languageWordService.getByLanguageAndKeyId(Messages.CarDeleted,Integer.parseInt(environment.getProperty("language"))));
-
+        return new SuccesResult(languageWordService.getByLanguageAndKeyId(Messages.CarDeleted));
     }
 
     @Override
@@ -82,7 +111,7 @@ public class CarManager implements CarService {
         List<CarSearchListDto> response = this.carDao.getByDailyPrice(dailyPrice).stream()
                 .map(car -> modelMapperService.forDto().map(car, CarSearchListDto.class)).collect(Collectors.toList());
 
-        return new SuccesDataResult<List<CarSearchListDto>>(response, languageWordService.getByLanguageAndKeyId(Messages.CarByDailyPriceListed,Integer.parseInt(environment.getProperty("language"))));
+        return new SuccesDataResult<List<CarSearchListDto>>(response, languageWordService.getByLanguageAndKeyId(Messages.CarByDailyPriceListed));
     }
 
     @Override
@@ -90,7 +119,7 @@ public class CarManager implements CarService {
 
         List<CarDetail> response = this.carDao.getCarWithBrandAndColorDetails();
 
-        return new SuccesDataResult<List<CarDetail>>(response, languageWordService.getByLanguageAndKeyId(Messages.CarDetailedListed,Integer.parseInt(environment.getProperty("language"))));
+        return new SuccesDataResult<List<CarDetail>>(response, languageWordService.getByLanguageAndKeyId(Messages.CarDetailedListed));
     }
 
     @Override
@@ -99,7 +128,7 @@ public class CarManager implements CarService {
                 .map(car -> modelMapperService.forDto().map(car, CarDetailForColorAndBrand.class))
                 .collect(Collectors.toList());
 
-        return new SuccesDataResult<List<CarDetailForColorAndBrand>>(response, languageWordService.getByLanguageAndKeyId(Messages.CarByColorIdListed,Integer.parseInt(environment.getProperty("language"))));
+        return new SuccesDataResult<List<CarDetailForColorAndBrand>>(response, languageWordService.getByLanguageAndKeyId(Messages.CarByColorIdListed));
     }
 
     @Override
@@ -108,7 +137,7 @@ public class CarManager implements CarService {
                 .map(car -> modelMapperService.forDto().map(car, CarDetailForColorAndBrand.class))
                 .collect(Collectors.toList());
 
-        return new SuccesDataResult<List<CarDetailForColorAndBrand>>(response, languageWordService.getByLanguageAndKeyId(Messages.CarByBrandIdListed,Integer.parseInt(environment.getProperty("language"))));
+        return new SuccesDataResult<List<CarDetailForColorAndBrand>>(response, languageWordService.getByLanguageAndKeyId(Messages.CarByBrandIdListed));
     }
 
     @Override
@@ -117,7 +146,7 @@ public class CarManager implements CarService {
                 .map(car -> modelMapperService.forDto().map(car, CarDetailForColorAndBrand.class))
                 .collect(Collectors.toList());
 
-        return new SuccesDataResult<List<CarDetailForColorAndBrand>>(response, languageWordService.getByLanguageAndKeyId(Messages.CarByCarIdListed,Integer.parseInt(environment.getProperty("language"))));
+        return new SuccesDataResult<List<CarDetailForColorAndBrand>>(response, languageWordService.getByLanguageAndKeyId(Messages.CarByCarIdListed));
     }
 
     @Override
@@ -130,7 +159,7 @@ public class CarManager implements CarService {
         List<CarSearchListDto> result = this.carDao.getAllWithoutMaintenanceOfCar();
         List<CarSearchListDto> response = result.stream()
                 .map(car -> modelMapperService.forDto().map(car, CarSearchListDto.class)).collect(Collectors.toList());
-        return new SuccesDataResult<List<CarSearchListDto>>(response, languageWordService.getByLanguageAndKeyId(Messages.CarWithoutMaintenanceListed,Integer.parseInt(environment.getProperty("language"))));
+        return new SuccesDataResult<List<CarSearchListDto>>(response, languageWordService.getByLanguageAndKeyId(Messages.CarWithoutMaintenanceListed));
     }
 
     @Override
@@ -138,13 +167,13 @@ public class CarManager implements CarService {
         List<CarSearchListDto> response = this.carDao.getByCity_CityId(cityId).stream()
                 .map(car -> modelMapperService.forDto().map(car, CarSearchListDto.class)).collect(Collectors.toList());
 
-        return new SuccesDataResult<List<CarSearchListDto>>(response, languageWordService.getByLanguageAndKeyId(Messages.CarByCityIdListed,Integer.parseInt(environment.getProperty("language"))));
+        return new SuccesDataResult<List<CarSearchListDto>>(response, languageWordService.getByLanguageAndKeyId(Messages.CarByCityIdListed));
     }
 
     @Override
     public Result checkCarExistsInGallery(int id) {
         if (!this.carDao.existsById(id)) {
-            return new ErrorResult(languageWordService.getByLanguageAndKeyId(Messages.CarNotFound,Integer.parseInt(environment.getProperty("language"))));
+            return new ErrorResult(languageWordService.getByLanguageAndKeyId(Messages.CarNotFound));
         }
         return new SuccesResult();
     }
